// src/app/model/mscustomer.ts
export class mscustomer {
    idcard: string;
    customerCode: string;
    title: string;
    fname: string;
    lname: string;
    fnameEng: string;
    lnameEng: string;
    nameOth: string;
    secondaryName: string;
    secondaryNameEng: string;
    image: string;
    birthday: string;
    race: string;
    religion: string;
    placeofBirth: string;
    otherContactInfo: string;
    occupation: string;
    sex: string;
    status: string;
    phone: string;
    passport: string;
    driver_card: string;
    otherContact: string;
    email: string;
    active: string;
    remark: string;
    createdBy: string;
    createDate: Date;
    updatedBy: string;
    updateDate: Date;
    sourceId: string;
    source: string;
};